+++ packages/react/src/interface.ts
@@ -1,6 +1,6 @@
-import { CuriConfig, AnyResponse } from '@curi/core';
+import { CuriConfig, Response } from '@curi/core';
 
 export interface CuriContext {
   curi: CuriConfig;
-  curiResponse: AnyResponse;
+  curiResponse: Response;
 }